Refactor common code.
